feat: it adds authentication on mobile and transfers the code to index.tsx
